jira-to-msproject/requirements.txt
requests>=2.28.1
python-dotenv>=0.21.0
pydantic>=1.10.2
pywin32>=305